refactor(cli): split access.go by concern (E4)
Carve the request-composition engine into access_plan.go, the *E run
methods into access_run.go, and the printers into access_print.go.
access.go keeps the cobra constructors and option types. Intra-package
moves only.
